+++ config/packages/messenger.yaml
@@ -4,7 +4,7 @@ framework:
             messenger.bus.default:
                 middleware:
                     - 'App\Middleware\ExceptionCatcherMiddleware'
-                    - 'validation'
+                    #- 'validation'
         # Uncomment this (and the failed transport below) to send failed messages to this transport for later handling.
         # failure_transport: failed
